[Persistence][15] Extract CoreData managed model calsses into separate file
